% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_pipe.R
\name{\%P>\%}
\alias{\%P>\%}
\title{\code{\%P>\%} Prints and pipes}
\usage{
lhs \%P>\% rhs
}
\arguments{
\item{lhs}{The left hand side of the pipe.}

\item{rhs}{The right hand side of the pipe}
}
\description{
This pipe variant prints the object received from the left hand side prior
  to piping it to the right hand side.
}
\note{
This code is experimental. Use at your own risk.
}
\examples{
  if (!require(dplyr)) stop("Examples need dplyr to run")
  mtcars \%>\%
    filter(am == 1) \%P>\%
    select(qsec)

}
