% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPipeChain.R
\name{printPipeChain}
\alias{printPipeChain}
\title{Prints each pipe step in current text selection}
\usage{
printPipeChain()
}
\description{
Reads the currently selected text from the RStudio API and prints
  for each pipestep the resulting object if unique. Data frames are
  converted by as.tibble(). Meant to be called as a RStudio addin.
}
