% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R, R/viewPipeChain.R
\name{viewPipeChain}
\alias{viewPipeChain}
\title{Creates a View() output for each pipe step in current text selection}
\usage{
viewPipeChain()

viewPipeChain()
}
\value{
No return value, called for side effects
}
\description{
Reads the currently selected text from the RStudio API and displays a data view
  in the source pane for each pipe step creating a unique object.
  Meant to be called as an RStudio addin.

Reads the currently selected text from the RStudio API and displays a data view
  in the source pane for each pipe step creating a unique object.
  Meant to be called as an RStudio addin.
}
