% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R, R/viewPipeChain.R
\name{printPipeChain}
\alias{printPipeChain}
\title{Prints each pipe step in current text selection}
\usage{
printPipeChain()

printPipeChain()
}
\value{
No return value, called for side effects
}
\description{
Reads the currently selected text from the RStudio API and prints
  for each pipe step the resulting object if unique. Data frames are
  converted by as_tibble(). Meant to be called as an RStudio addin.

Reads the currently selected text from the RStudio API and prints
  for each pipe step the resulting object if unique. Data frames are
  converted by as.tibble(). Meant to be called as an RStudio addin.
}
